#12 move SubNav component
